+++ FireDevil/FollowerTask_Scarecrow.cs
@@ -155,5 +155,7 @@ namespace FireDevil
                 //ProgressTask();
             }
         }
+
+        //OnArrive?
     }
 }