+++ FireDevil/changelog.md
@@ -1,10 +1,19 @@
 # Changelog
 
+## [1.0.2]
+- improved menu
+- fixed issue that would lock the cursor when using the mod manager
+- added option to disable most fleece penalties
+- added option to increase shrine/outhouse/silo storage
+- added option to have follower loyalty overflow to next level
+- added option to instantly pickup mine resources
+- added menu for follower traits
+
 ## [1.0.1]
 - added mod menu
-- added mines do not exhaused
+- added mines do not exhaust
 - added job description in follower thoughts
-- added options to toggle doctines
+- added options to toggle doctrines
 - added weights for randomized weapons; resets on restart
 
 ## [1.0.0]